multiagent_system/requirements.txt
# Основные зависимости
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
statsmodels>=0.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# LLM интеграция
openai>=1.3.0
anthropic>=0.8.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.10

# Утилиты
pydantic>=2.5.0
pyyaml>=6.0
joblib>=1.3.0
tqdm>=4.65.0
python-dotenv>=1.0.0
redis>=5.0.0

# Тестирование
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.6.0

# Документация
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.0.0